configurator#1 実装

+++ include/seed_smartactuator_sdk/aero3_command.h
@@ -98,8 +98,8 @@ namespace aero
       void runScript(uint8_t _number,uint16_t _data);
       void setControllerCmd();
 
-      void write_1byte(uint16_t _address, int8_t *_write_data);
-      void write_2byte(uint16_t _address, int8_t *_write_data);
+      void write_1byte(uint16_t _address, uint8_t *_write_data);
+      void write_2byte(uint16_t _address, uint16_t *_write_data);
       std::vector<uint8_t> read_1byte(uint16_t _address);
       std::vector<uint8_t> read_2byte(uint16_t _address);
